Ceated a Header comp with atomic structure
